Fix list_display handling.
The django documentation for list_display lists four ways of specifying
elements: field names, callables, admin methods and model methods.

The current code in templatetags/mptt-admin.py takes only field names and
model methods into account: When the element is a callable, it throws a
ValueError in getattr(), when it is an admin method, it ignores it.

This commit fixes the issue by just using any field_name except
'action_checkbox' as mptt_indent_field.