Have Travis CI use the tox configuration.
This reduces maintenance and allows contributors to test in the same way
as the CI system. The .travis.yml is simpler as a result.

The tox configuration needed repair - it was failing for all virtual
environments.

Removed the now unnecessary shell script & requirements.txt files by
moving requirements to setup.py and tox.ini files.