Update upgrade.rst to include 0.9.1 changelog
Hi, this changelog PR may be considered a work in progress. I haven't work on any of the changes and I just got the changes from the commit messages. Not sure if everything is relevant. I did my best but I'm pretty sure you could make it better. Hope that helps. Cheers !